server/requirements.txt
# LLM & Structured Parsing
instructor==1.7.0
pydantic==2.10.6
openai==1.59.8

# Web Scraping
scrapling==0.2.1
trafilatura==1.12.2  # HTML → clean markdown for enrichment/crawler.py
python-jobspy==1.1.82  # Tier 1 job board scraper (scraper/jobspy_run.py)

# Database
psycopg2-binary==2.9.10

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0  # Retry logic for resilient scraping
structlog>=24.4.0

# Gmail API (for reply detection)
google-auth-oauthlib==1.2.0
google-api-python-client==2.137.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Secret Management
keyring==25.5.0
cryptography==43.0.1  # AES-256-GCM for crypto_helper.py (decrypts user_config secrets at read time)

# Analytics (Optional - for skill extraction from free text)
# keybert==0.8.5
# sentence-transformers==2.2.2
# Uncomment if using KeyBERT instead of structured JSONB input
